refactor(server): extract operator-console handlers (ADR 0023 phase 3)
Finishes the half-done operator_api/ extraction the ADR called out: the 21
React-console handler bodies (runtime-status, subagent, scheduler, goal,
workflow, activity, inbox, chat-commands) were inline closures in _main that
closed over the once-ambient globals. They move verbatim into
operator_api/console_handlers.py as module-level functions (deps imported under
the same alias names so bodies are byte-identical; the one captured local, the
operator project root, resolves live). _main imports the module and passes the
functions to register_operator_routes instead of defining 21 closures.

server/__init__.py drops ~300 lines (1106 -> 806). Adds
tests/test_console_handlers.py (inbox-auth open/match, store-off degradation
paths, chat-command listing). No change to register_operator_routes or its tests
(they pass their own callbacks).
